datebase updated correctly now gui does not build

+++ EveMailHelper.ServiceLayer/Interfaces/ITaxManager.cs
@@ -2,13 +2,13 @@
 {
     public interface ITaxManager
     {
-        double AddBrokerTax(double isk, double BrokerRelationsLevel, double FactionStanding, double CorpStanding);
+        double AddBrokerTax(double isk, int BrokerRelationsLevel, double FactionStanding, double CorpStanding);
         double AddSalesTax(double isk, int AccountSkillLevel);
-        double CalculateBrokerTax(double isk, double BrokerRelationsLevel, double FactionStanding, double CorpStanding);
-        double CalculateImmediateSellTaxes(double isk, double BrokerRelationsLevel, double FactionStanding, double CorpStanding);
+        double CalculateBrokerTax(double isk, int BrokerRelationsLevel, double FactionStanding, double CorpStanding);
+        double CalculateImmediateSellTaxes(double isk, int BrokerRelationsLevel, double FactionStanding, double CorpStanding);
         double CalculateSalesTax(double isk, int AccountSkillLevel);
-        double CalculateSellOrderTaxes(double isk, int AccountSkillLevel, double BrokerRelationsLevel, double FactionStanding, double CorpStanding);
-        double GetBrokerTaxPercent(double BrokerRelationsLevel, double FactionStanding, double CorpStanding);
+        double CalculateSellOrderTaxes(double isk, int AccountSkillLevel, int BrokerRelationsLevel, double FactionStanding, double CorpStanding);
+        double GetBrokerTaxPercent(int BrokerRelationsLevel, double FactionStanding, double CorpStanding);
         double GetSalesTaxPercent(int AccountSkillLevel);
     }
 }
